overkill Makefile and a real header lol

+++ global.c
@@ -1,4 +1,4 @@
 #include "geodesics.h"
 
-long double (*geodesic_points)[4] = NULL;
+double (*geodesic_points)[4] = NULL;
 int num_points = 0;
